refactor(Constraint Authoring): Move helper functions to new class
#1231
