+++ src/assets/wise5/authoringTool/constraint/edit-constraint-removal-criteria/edit-constraint-removal-criteria.component.ts
@@ -4,6 +4,7 @@ import { RemovalCriteria } from '../../../../../app/domain/removalCriteria';
 import { RemovalCriteriaParam } from '../../../../../app/domain/removalCriteriaParam';
 import { ComponentContent } from '../../../common/ComponentContent';
 import { MultipleChoiceContent } from '../../../components/multipleChoice/MultipleChoiceContent';
+import { EditConstraintRemovalCriteriaHelper } from './edit-constraint-removal-criteria-helper';
 
 @Component({
   selector: 'edit-constraint-removal-criteria',
@@ -19,15 +20,9 @@ export class EditConstraintRemovalCriteriaComponent implements OnInit {
   );
   @Input() constraint: any;
   @Input() criteria: any;
-  criteriaNameToComponentType: Map<string, string> = new Map(
-    Object.entries({
-      choiceChosen: 'MultipleChoice',
-      fillXNumberOfRows: 'Table',
-      wroteXNumberOfWords: 'OpenResponse'
-    })
-  );
   @Input() node: any;
   nodeIds: string[];
+  private removalCriteriaHelper: EditConstraintRemovalCriteriaHelper;
   private stepParam: RemovalCriteriaParam = new RemovalCriteriaParam('nodeId', $localize`Step`);
 
   protected removalCriteria = [
@@ -90,15 +85,19 @@ export class EditConstraintRemovalCriteriaComponent implements OnInit {
 
   ngOnInit(): void {
     this.allNodeIds = this.projectService.getFlattenedProjectAsNodeIds(true);
+    this.removalCriteriaHelper = new EditConstraintRemovalCriteriaHelper(
+      this.projectService,
+      this.componentIdToIsSelectable
+    );
     this.setNodeIds();
-    this.calculateSelectableComponents(this.criteria);
+    this.removalCriteriaHelper.calculateSelectableComponents(this.criteria);
   }
 
   private setNodeIds(): void {
-    this.nodeIds = this.criteriaNameToComponentType.has(this.criteria.name)
+    this.nodeIds = this.removalCriteriaHelper.hasCriteriaNameToComponentType(this.criteria.name)
       ? this.getStepsWithComponentType(
           this.allNodeIds,
-          this.criteriaNameToComponentType.get(this.criteria.name)
+          this.removalCriteriaHelper.getCriteriaNameToComponentType(this.criteria.name)
         )
       : this.allNodeIds;
   }
@@ -125,13 +124,16 @@ export class EditConstraintRemovalCriteriaComponent implements OnInit {
     for (const paramObject of params) {
       const value = paramObject.value;
       criteria.params[value] = paramObject.defaultValue;
-      if (value === 'nodeId' && this.stepContainsAcceptableComponent(this.node.id, criteria)) {
+      if (
+        value === 'nodeId' &&
+        this.removalCriteriaHelper.stepContainsAcceptableComponent(this.node.id, criteria)
+      ) {
         criteria.params[value] = this.node.id;
       }
     }
     this.setNodeIds();
-    this.calculateSelectableComponents(criteria);
-    this.automaticallySelectComponentIfPossible(criteria);
+    this.removalCriteriaHelper.calculateSelectableComponents(criteria);
+    this.removalCriteriaHelper.automaticallySelectComponentIfPossible(criteria);
     this.saveProject();
   }
 
@@ -144,69 +146,11 @@ export class EditConstraintRemovalCriteriaComponent implements OnInit {
     return [];
   }
 
-  private stepContainsAcceptableComponent(nodeId: string, criteria: any): boolean {
-    if (this.criteriaNameToComponentType.has(criteria.name)) {
-      return this.hasComponentType(nodeId, this.criteriaNameToComponentType.get(criteria.name));
-    }
-    return true;
-  }
-
-  private hasComponentType(nodeId: string, componentType: string): boolean {
-    return (
-      this.projectService
-        .getComponents(nodeId)
-        .filter((component) => component.type === componentType).length > 0
-    );
-  }
-
   protected nodeIdChanged(criteria: any): void {
     criteria.params.componentId = '';
-    this.calculateSelectableComponents(criteria);
-    this.automaticallySelectComponentIfPossible(criteria);
+    this.removalCriteriaHelper.calculateSelectableComponents(criteria);
+    this.removalCriteriaHelper.automaticallySelectComponentIfPossible(criteria);
     this.saveProject();
-  }
-
-  private calculateSelectableComponents(criteria: any): void {
-    const components = this.projectService.getComponents(criteria.params.nodeId);
-    if (this.criteriaNameToComponentType.has(criteria.name)) {
-      this.setSelectableComponents(components, this.criteriaNameToComponentType.get(criteria.name));
-    } else {
-      this.makeAllComponentsSelectable(components);
-    }
-  }
-
-  private setSelectableComponents(components: ComponentContent[], componentType: string): void {
-    components.forEach((component) => {
-      this.componentIdToIsSelectable[component.id] = component.type === componentType;
-    });
-  }
-
-  private makeAllComponentsSelectable(components: ComponentContent[]): void {
-    components.forEach((component) => {
-      this.componentIdToIsSelectable[component.id] = true;
-    });
-  }
-
-  private automaticallySelectComponentIfPossible(criteria: any): void {
-    if (this.criteriaNameToComponentType.has(criteria.name)) {
-      this.selectIfOnlyOneOfComponentType(
-        criteria,
-        this.projectService.getComponents(criteria.params.nodeId),
-        this.criteriaNameToComponentType.get(criteria.name)
-      );
-    }
-  }
-
-  private selectIfOnlyOneOfComponentType(
-    criteria: any,
-    components: ComponentContent[],
-    componentType: string
-  ): void {
-    if (components.filter((component) => component.type === componentType).length === 1) {
-      criteria.params.componentId = components.find(
-        (component) => component.type === componentType
-      ).id;
-    }
   }
 
   protected getNodePosition(nodeId: string): string {
